+++ src/system/rocket_system.cpp
@@ -141,7 +141,7 @@ void RocketSystem::on(const protocol::message::set_arm_state_message_t& m) {
 void RocketSystem::updateStreams(SensorMeasurements meas, WorldEstimate est, ActuatorSetpoint& sp) {
   protocol::message::system_message_t m {
     .time = ST2MS(chibios_rt::System::getTime()),
-    .state = (int) state,
+    .state = (uint8_t) state,
     .motorDC = sp.throttle
   };
   logger.write(m);
